# slides/se332/openapi.yaml
...
openapi: 3.0.0
info:
  title: UIT Course & Student Manager API
  version: 1.0.0
  description: Comprehensive CRUD API powered by Redis Stack (RedisJSON + RediSearch)
...
paths:
...
  /students:
    post:
      summary: Create a new student (JSON.SET)
      tags: [student]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentCreate'
      responses:
        201:
          description: Student created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
    get:
      summary: List all students with pagination
      tags: [student]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        200:
          description: List of students
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedStudents'
  /students/{id}:
    get:
      summary: Get student by ID (JSON.GET)
      tags: [student]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Student found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        404:
          description: Student not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: Partially update student (JSON.SET on paths)
      tags: [student]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentUpdate'
      responses:
        200:
          description: Student updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        404:
          description: Student not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Replace entire student record
      tags: [student]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentCreate'
      responses:
        200:
          description: Student replaced
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        404:
          description: Student not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete student by ID
      tags: [student]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        204:
          description: Student deleted
        404:
          description: Student not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /students/search:
    get:
      summary: Search students (FT.SEARCH with filters)
      tags: [student]
      parameters:
        - name: query
          in: query
          schema:
            type: string
          description: Full-text search (name, username)
        - name: cohort
          in: query
          schema:
            type: string
          description: Filter by cohort
        - name: minGpa
          in: query
          schema:
            type: number
          description: Minimum GPA
        - name: maxGpa
          in: query
          schema:
            type: number
          description: Maximum GPA
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [gpa, name, cohort]
            default: gpa
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        200:
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentSearchResult'
...
  /courses/{courseId}/students:
    post:
      summary: Enroll student in course (JSON.SET + XADD)
      description: Updates Course/Student JSON documents and appends an enrollment stream event
      tags: [course]
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                studentId:
                  type: string
      responses:
        201:
          description: Student enrolled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentResult'
    get:
      summary: Get all enrolled students in course
      tags: [course]
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: List of enrolled students
          content:
            application/json:
              schema:
                type: object
                properties:
                  courseId:
                    type: string
                    example: SE332.Q21
                  students:
                    type: array
                    items:
                      $ref: '#/components/schemas/Student'
                  total:
                    type: integer
                    example: 3
  /courses/{courseId}/students/{studentId}:
    delete:
      summary: Unenroll student from course (JSON.SET + XADD)
      description: Updates Course/Student JSON documents and appends an unenrollment stream event
      tags: [course]
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
        - name: studentId
          in: path
          required: true
          schema:
            type: string
      responses:
        204:
          description: Student unenrolled
        404:
          description: Course or student not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: UUID
  schemas:
...
    Student:
      type: object
      properties:
        id:
          type: string
          example: 23521476
        username:
          type: string
          example: thiendp
        name:
          type: string
          example: Dang Phu Thien
        cohort:
          type: string
          example: 23
        gpa:
          type: number
          example: 9.2
        enrollments:
          type: array
          items:
            type: string
          example: [SE332.Q21, SE121.Q21]
    StudentCreate:
      type: object
      required:
        - id
        - username
        - name
        - cohort
        - gpa
      properties:
        id:
          type: string
        username:
          type: string
        name:
          type: string
        cohort:
          type: string
        gpa:
          type: number
    StudentUpdate:
      type: object
      properties:
        username:
          type: string
        name:
          type: string
        cohort:
          type: string
        gpa:
          type: number
    PaginatedStudents:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Student'
        total:
          type: integer
          description: Total number of students matching the query
          example: 42
        limit:
          type: integer
          example: 10
        offset:
          type: integer
          example: 0
    StudentSearchResult:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Student'
        total:
          type: integer
          description: Total number of matching students
          example: 5
        query:
          type: string
          description: The search query used
          example: thien
...
    EnrollmentResult:
      type: object
      properties:
        courseId:
          type: string
          example: SE332.Q21
        studentId:
          type: string
          example: 23521476
        enrolledAt:
          type: string
          format: date-time
          example: 2026-03-14T10:00:00Z
        streamId:
          type: string
          description: Redis Stream entry ID from XADD
          example: 1741950000000-0
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Not found
        message:
          type: string
          example: Student with id 23521476 not found